+++ docs/conf.py
@@ -215,8 +215,6 @@ man_pages = [
      [u'Jannis Leidel'], 1)
 ]
 
-
-# Example configuration for intersphinx: refer to the Python standard library.
 # Example configuration for intersphinx: refer to the Python standard library.
 intersphinx_mapping = {
     'python': ('http://python.readthedocs.org/en/v2.7.2/', None),